requirements.txt
numpy
pandas
opencv-python
Pillow
matplotlib
scikit-image